Add support for OG image and Twitter cards
Add the frontmatter fields 'image', a relative path to an image,
and 'imageAlt', where 'image' will be handled by gatsby
filesystem and Sharp. Add a default, fallback OG image and alt
text. The OG and Twitter metadata is added via React Helmet in
the SEO component. If an image is provided it is assumed to be
at least 1200 by 600 px and therefore the larger
'summary_large_image' Twitter card type is used. If no image is
provided and the default OG image, a simple icon, is used, the
smaller Twitter card type, 'summary', is used. Add an image for
one post.